[CHIPCRAFT][TOOLCHAIN][BUILDSYSTEM] Detect whether ct-ng was already built.
If given ct-ng source directory contains built ct-ng binary, it is
assumed the project was already built. In that case makefile only
installs the built sources to known location.

This change will be used in the future in Jenkins artifacts.
chipcraft-toolchain project will be given a prebuilt
chipcraft-crosstool-ng project, which will shorten the build times.