[CHIPCRAFT][TOOLCHAIN][BUILDSYSTEM] Better checks for git, libtool. Added yacc check.
Yacc, or something able to be used like yacc, is needed by ct-ng.
Unfortunately, the standard check will set YACC to 'yacc' by default,
even if it's not found. Therefore we will check that ${YACC} actually
exists and can be executed. We do not want to fail during ct-ng build.